updated js

+++ a10-bank-app/assets/js/main.js
@@ -173,7 +173,12 @@ doc_ready(() => {
           (userIndex) => userIndex.accountNumber == to_user
         );
 
-      if (users[senderCheck] == null || users[senderCheck] == "") {
+      if (
+        (users[senderCheck] == null || users[senderCheck] == "") &&
+        (users[receiverCheck] == null || users[receiverCheck] == "")
+      ) {
+        alert("Users not found!");
+      } else if (users[senderCheck] == null || users[senderCheck] == "") {
         alert("Sender's account not found!");
       } else if (users[receiverCheck] == null || users[receiverCheck] == "") {
         alert("Receiver's account not found!");
@@ -557,17 +562,22 @@ doc_ready(() => {
 
   // PROMPT FOR CLEARING ALL DATA, TO PREVENT ACCIDENTAL DELETION
   add_event(id("clear-all-btn"), "click", () => {
-    let clearPrompt = prompt(
-        'Are you sure to delete all stored datas?\nType "Y" for yes and "N" for no.',
-        "N"
-      ),
-      clearAnswer = clearPrompt.toLowerCase();
+    let users = FnHandler.userStorage();
 
-    if (clearAnswer == "y") {
-      window.localStorage.clear();
-      FnHandler.list_users();
-    } else {
-      return;
+    if (users.length != 0) {
+      let clearPrompt = prompt(
+          'Are you sure to delete all stored datas?\nType "Y" for yes and "N" for no.',
+          "N"
+        ),
+        clearAnswer =
+          clearPrompt != null ? clearPrompt.toLowerCase() : console.clear();
+
+      if (clearAnswer == "y") {
+        window.localStorage.clear();
+        FnHandler.list_users();
+      } else {
+        return;
+      }
     }
   });
